2021/CVE-2021-22054/requirements.txt
pycryptodome>=3.14.1
requests>=2.25.1
